Use off_t and size_t types for input functions.
git-svn-id: https://smlnj-gforge.cs.uchicago.edu/svn/smlnj@5898 d172b47e-f5d9-4df3-ab46-1c4da947e531

+++ sml/trunk/runtime/gc/heap-input.h
@@ -21,7 +21,7 @@ typedef struct {	    /* An input source for reading heap data.  We need */
 				/* buffered characters are exhausted */
     Byte_t	*base;		/* the start of the bufferec characters */
     Byte_t	*buf;		/* the current position in the buffer */
-    long	nbytes;
+    size_t	nbytes;
 } inbuf_t;
 
 
@@ -45,7 +45,7 @@ typedef struct {	/* big-object region relocation info */
 
 /* Utility routines */
 extern ml_val_t *HeapIO_ReadExterns (inbuf_t *bp);
-extern status_t HeapIO_Seek (inbuf_t *bp, long offset);
-extern status_t HeapIO_ReadBlock (inbuf_t *bp, void *blk, long len);
+extern status_t HeapIO_Seek (inbuf_t *bp, off_t offset);
+extern status_t HeapIO_ReadBlock (inbuf_t *bp, void *blk, off_t len);
 
 #endif /* !_HEAP_INPUT_ */
